Add files via upload
This programm was written to facilitate the management of creation of 500+ similar, but different documents. It is based on a storage and naming system of documents, which has to be agrred in advance. Basically, it first checks if some document is present in more then one folder and then it gives an overview, of how many and which exactly documents are in every folder in the directory tree.