Add files via upload
This programm automates the 'happy birthday' emailing. It checks for any birthday children from an excel file, if there're none - it send a corresponding email to chosen recepients; whereas if there're any - a gratulation email is created according to the gender of the birthday child and a picture is randomly chosen from an images' directory. Then an email goes out to the birthday child and a predefined list of recipients.